Finance/requirements.txt
google-api-python-client==1.9.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.6.0
httplib2==0.18.1
oauth2client==4.1.3
bs4
pandas_datareader
plotly
tushare
mplfinance~=0.12.6a3
OpenDataTools~=1.0.0


